service/requirements.txt
altgraph==0.17.4
aniso8601==9.0.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
certifi==2021.5.30
cffi==1.16.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
daphne==3.0.2
distlib==0.3.8
Django==4.0
djangorestframework==3.15.1
filelock==3.13.1
Flask==3.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.26.4
packaging==24.0
panda==0.3.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.0
psutil==5.9.8
py_vncorenlp==0.1.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygame==2.5.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.4
pyjnius==1.6.1
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.25.1
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.28
sqlparse==0.5.0
tk==0.1.0
Twisted==24.3.0
# twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.6
virtualenv==20.25.1
websocket-client==1.8.0
Werkzeug==3.0.1
whitenoise==6.6.0
zope.interface==6.3
